warnings removed

+++ TerminalGameEngine/GameObject.h
@@ -36,8 +36,8 @@ public:
 	GameObject(int xPos, int yPos);
 	int GetPosX() const { return xPos; }
 	int GetPosY() const { return yPos; }
-	int GetMaxPosX() const { return xPos + GetModelWidth() - 1; }
-	int GetMaxPosY() const { return yPos + GetModelHeight() - 1; }
+	int GetMaxPosX() const { return xPos + static_cast<int>(GetModelWidth()) - 1; }
+	int GetMaxPosY() const { return yPos + static_cast<int>(GetModelHeight()) - 1; }
 	size_t GetModelWidth() const;
 	size_t GetModelHeight() const { return model.size();}
 	const std::vector<std::vector<char>>& GetModel();
